+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @qrgrid/core
 
+## 0.2.0
+
+### Minor Changes
+
+- QR property name change noOfModules -> gridSize
+
 ## 0.1.2
 
 ### Patch Changes