+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @qrgrid/vue
 
+## 1.3.0
+
+### Minor Changes
+
+- QR property name change noOfModules -> gridSize
+
+### Patch Changes
+
+- Updated dependencies
+  - @qrgrid/core@0.2.0
+
 ## 1.2.2
 
 ### Patch Changes